requirements.txt
PyQt5
pyserial
pygame
